# [PERF] base,web: improve (web_)load_menus performance