--- a/addons/web/tests/test_perf_load_menu.py
+++ b/addons/web/tests/test_perf_load_menu.py
@@ -18,15 +18,15 @@ class TestPerfSessionInfo(common.HttpCase):
         self.authenticate(user.login, "info")
 
         self.env.registry.clear_all_caches()
-        # cold ormcache (only web: 42, all module: 131)
-        with self.assertQueryCount(131):
+        # cold ormcache (only web: 42, all module: 112)
+        with self.assertQueryCount(112):
             self.url_open(
                 "/web/session/get_session_info",
                 data=json.dumps({'jsonrpc': "2.0", 'method': "call", 'id': str(uuid4())}),
                 headers={"Content-Type": "application/json"},
             )
 
-        # cold fields cache - warm ormcache (only web: 7, all module: 22)
+        # cold fields cache - warm ormcache (only web: 6, all module: 22)
         with self.assertQueryCount(22):
             self.url_open(
                 "/web/session/get_session_info",
@@ -37,20 +37,20 @@ class TestPerfSessionInfo(common.HttpCase):
     def test_load_web_menus_perf(self):
         self.env.registry.clear_all_caches()
         self.env.invalidate_all()
-        # cold orm/fields cache (only web: 19, all module: 80)
-        with self.assertQueryCount(80):
+        # cold orm/fields cache (only web: 10, all module: 38)
+        with self.assertQueryCount(38):
             self.env['ir.ui.menu'].load_web_menus(False)
 
-        # cold fields cache - warm orm cache (only web: 6, all module: 635)
+        # cold fields cache - warm orm cache (only web: 0, all module: 1)
         self.env.invalidate_all()
-        with self.assertQueryCount(635):
+        with self.assertQueryCount(1):
             self.env['ir.ui.menu'].load_web_menus(False)
 
     def test_load_menus_perf(self):
         self.env.registry.clear_all_caches()
         self.env.invalidate_all()
-        # cold orm/fields cache (only web: 12, all module: 45)
-        with self.assertQueryCount(45):
+        # cold orm/fields cache (only web: 10, all module: 38)
+        with self.assertQueryCount(38):
             self.env['ir.ui.menu'].load_menus(False)
 
         # cold fields cache - warm orm cache (only web: 0, all module: 1)
@@ -61,11 +61,11 @@ class TestPerfSessionInfo(common.HttpCase):
     def test_visible_menu_ids(self):
         self.env.registry.clear_all_caches()
         self.env.invalidate_all()
-        # cold ormcache (only web: 5, all module: 19)
-        with self.assertQueryCount(19):
+        # cold ormcache (only web: 5, all module: 14)
+        with self.assertQueryCount(14):
             self.env['ir.ui.menu']._visible_menu_ids()
 
-        # cold fields cache - warm orm cache (only web: 1, all module: 1)
+        # cold fields cache - warm orm cache (only web: 0, all module: 0)
         self.env.invalidate_all()
-        with self.assertQueryCount(1):
+        with self.assertQueryCount(0):
             self.env['ir.ui.menu']._visible_menu_ids()