--- a/odoo/addons/base/models/ir_ui_menu.py
+++ b/odoo/addons/base/models/ir_ui_menu.py
@@ -234,92 +234,89 @@ class IrUiMenu(models.Model):
     @api.model
     @tools.ormcache_context('self._uid', 'debug', keys=('lang',))
     def load_menus(self, debug):
-        """ Loads all menu items (all applications and their sub-menus).
-
-        :return: the menu root
-        :rtype: dict('children': menu_nodes)
-        """
-        fields = ['name', 'sequence', 'parent_id', 'action', 'web_icon']
-        menu_roots = self.get_user_roots()
-        menu_roots_data = menu_roots.read(fields) if menu_roots else []
-        menu_root = {
-            'id': False,
-            'name': 'root',
-            'parent_id': [-1, ''],
-            'children': [menu['id'] for menu in menu_roots_data],
-        }
-
-        all_menus = {'root': menu_root}
-
-        if not menu_roots_data:
-            return all_menus
-
-        # menus are loaded fully unlike a regular tree view, cause there are a
-        # limited number of items (752 when all 6.1 addons are installed)
-        menus_domain = [('id', 'child_of', menu_roots.ids)]
         blacklisted_menu_ids = self._load_menus_blacklist()
-        if blacklisted_menu_ids:
-            menus_domain = expression.AND([menus_domain, [('id', 'not in', blacklisted_menu_ids)]])
-        menus = self.search(menus_domain)._filter_visible_menus()
-        menu_items = menus.read(fields)
-        xmlids = (menu_roots + menus)._get_menuitems_xmlids()
+        visible_menus = self.search_fetch(
+            [('id', 'not in', blacklisted_menu_ids)],
+            ['name', 'parent_id', 'action', 'web_icon'],
+        )._filter_visible_menus()
 
-        # add roots at the end of the sequence, so that they will overwrite
-        # equivalent menu items from full menu read when put into id:item
-        # mapping, resulting in children being correctly set on the roots.
-        menu_items.extend(menu_roots_data)
+        children_dict = defaultdict(list)  # {parent_id: []} / parent_id == False for root menus
+        for menu in visible_menus:
+            children_dict[menu.parent_id.id].append(menu.id)
 
-        mi_attachments = self.env['ir.attachment'].sudo().search_read(
+        app_info = {}
+        # recursively set app ids to related children
+        def _set_app_id(menu_app_id, menu_id):
+            app_info[menu_id] = menu_app_id
+            for child_id in children_dict[menu_id]:
+                _set_app_id(menu_app_id, child_id)
+
+        for root_menu_id in children_dict[False]:
+            _set_app_id(root_menu_id, root_menu_id)
+
+        # Filter out menus not related to an app (+ keep root menu), it happens when
+        # some parent menu are not visible for group.
+        visible_menus = visible_menus.filtered(lambda menu: menu.id in app_info)
+
+        xmlids = visible_menus._get_menuitems_xmlids()
+        icon_attachments = self.env['ir.attachment'].sudo().search_read(
             domain=[('res_model', '=', 'ir.ui.menu'),
-                    ('res_id', 'in', [menu_item['id'] for menu_item in menu_items if menu_item['id']]),
+                    ('res_id', 'in', visible_menus._ids),
                     ('res_field', '=', 'web_icon_data')],
             fields=['res_id', 'datas', 'mimetype'])
+        icon_attachments_res_id = {attachment['res_id']: attachment for attachment in icon_attachments}
 
-        mi_attachment_by_res_id = {attachment['res_id']: attachment for attachment in mi_attachments}
+        menus_dict = {}
+        action_ids_by_type = defaultdict(list)
+        for menu in visible_menus:
 
-        # set children ids and xmlids
-        menu_items_map = {menu_item["id"]: menu_item for menu_item in menu_items}
-        for menu_item in menu_items:
-            menu_item.setdefault('children', [])
-            parent = menu_item['parent_id'] and menu_item['parent_id'][0]
-            menu_item['xmlid'] = xmlids.get(menu_item['id'], "")
-            if parent in menu_items_map:
-                menu_items_map[parent].setdefault(
-                    'children', []).append(menu_item['id'])
-            attachment = mi_attachment_by_res_id.get(menu_item['id'])
-            if attachment:
-                menu_item['web_icon_data'] = attachment['datas'].decode()
-                menu_item['web_icon_data_mimetype'] = attachment['mimetype']
+            menu_id = menu.id
+            attachment = icon_attachments_res_id.get(menu_id)
+
+            if action := menu.action:
+                action_model = action._name
+                action_id = action.id
+                action_ids_by_type[action_model].append(action_id)
             else:
-                menu_item['web_icon_data'] = False
-                menu_item['web_icon_data_mimetype'] = False
-        all_menus.update(menu_items_map)
+                action_model = False
+                action_id = False
 
-        # sort by sequence
-        for menu_id in all_menus:
-            all_menus[menu_id]['children'].sort(key=lambda id: all_menus[id]['sequence'])
+            menus_dict[menu_id] = {
+                'id': menu_id,
+                'name': menu.name,
+                'app_id': app_info[menu_id],
+                'action_model': action_model,
+                'action_id': action_id,
+                'web_icon': menu.web_icon,
+                'web_icon_data': attachment['datas'].decode() if attachment else False,
+                'web_icon_data_mimetype': attachment['mimetype'] if attachment else False,
+                'xmlid': xmlids.get(menu_id, ""),
+            }
 
-        # recursively set app ids to related children
-        def _set_app_id(app_id, menu):
-            menu['app_id'] = app_id
-            for child_id in menu['children']:
-                _set_app_id(app_id, all_menus[child_id])
+        # prefetch action.path
+        for model_name, action_ids in action_ids_by_type.items():
+            self.env[model_name].sudo().browse(action_ids).fetch(['path'])
 
-        for app in menu_roots_data:
-            app_id = app['id']
-            _set_app_id(app_id, all_menus[app_id])
+        # set children + model_path
+        for menu_dict in menus_dict.values():
+            if menu_dict['action_model']:
+                menu_dict['action_path'] = self.env[menu_dict['action_model']].sudo().browse(menu_dict['action_id']).path
+            else:
+                menu_dict['action_path'] = False
+            menu_dict['children'] = children_dict[menu_dict['id']]
 
-        # filter out menus not related to an app (+ keep root menu)
-        all_menus = {menu['id']: menu for menu in all_menus.values() if menu.get('app_id')}
-        all_menus['root'] = menu_root
-
-        return all_menus
+        menus_dict['root'] = {
+            'id': False,
+            'name': 'root',
+            'children': children_dict[False],
+        }
+        return menus_dict
 
     def _get_menuitems_xmlids(self):
-        menuitems = self.env['ir.model.data'].sudo().search([
-                ('res_id', 'in', self.ids),
-                ('model', '=', 'ir.ui.menu')
-            ])
+        menuitems = self.env['ir.model.data'].sudo().search_fetch(
+            [('res_id', 'in', self.ids), ('model', '=', 'ir.ui.menu')],
+            ['res_id', 'complete_name'],
+        )
 
         return {
             menu.res_id: menu.complete_name