## Rationals

The first request to the Odoo backend has to compute the `session_info` (ir.http) of the current user. 
In this calculation, most of the time is used to compute `load_menus`, which is cached, but not with a good hit rate, since it depends on the current user. The calculation of `load_menus` in `session_info` is only done to generate a hash to the webclient which is used for caching purposes (see `web_load_menus`).

The purpose of this PR is to improve and refactor the code of `load_menus` to be faster when the ormcache is cold (due to the low hit rate). A second PR will follow to avoid calling `load_menus` to create the cache key for the webclient in the first place (it will mostly add a new ormcache key attached to a SQL sequence that is sent back to the webclient and shared for menus and actions).

The speedscope didn't/doesn't a specific bottleneck but a lot of small ones. It is why we refactor `_visible_menu_ids` and `load_menus` completely for sake of performance and simplicity.

## Commits

### [IMP] web: add assertQueryCount for methods involved with load_menus

### [IMP] web: improve test_loaded_menu

For the next commit, it is important to have a realistic behavior and avoid patching the `search` method without calling the original one, as the current implementation only works with the override of `search_fetch`.

### [REF] base: refactor ir.ui.menu search_fetch()

The `search_fetch` of ir.ui.menu is overridden by default to filter the visible menu for the current user by default. It's not standard, and adds complexity to disable this behavior in many places.

### [REF] base: improve performance of `_visible_menu_ids`

Refactor and improve the performance of `_visible_menu_ids` to speed up load_menus() when the orm cache is cold.

Performance test:
```
from odoo.tools import mute_logger

def visible_menu():
    with mute_logger('odoo.registry'):
        self.env.registry.clear_all_caches()
    self.env.invalidate_all()
    self.env['ir.ui.menu']._visible_menu_ids()

%timeit visible_menu()
```
On the runbot database:
Before:
25.5 ms ± 203 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
After:
11 ms ± 148 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)


### [REF] base: refactor load_menus

Refactor `load_menus` for sake of performance and simplicity. Remove unused keys (parent_id, sequence and action), add action_path, action_id, action_model to avoid the very inefficient calculation of it in `load_web_menus`. In fact, `load_web_menus` doesn't set a prefetch set and then read the action path from the database one by one.
This makes `load_web_menus` less efficient when the ormcache is hot than when the ormcache is cold (because all path actions are already fetched from `_visible_menus_ids`, see changes no query count).

Note that one idea was to change the ormcache to depend on the list of the user groups, but `_load_menus_blacklist` depends on the user data (company_ids/employee_ids) inside some overrides. Also, the next step is to have a cache on the user's browser side, then, this ormcache might be useless in the future. Then, for the sake of simplicity, keep it that way for now.

Performance test:
```python
from odoo.tools import mute_logger

def load_menus():
    with mute_logger('odoo.registry'):
        self.env.registry.clear_all_caches()
    self.env.invalidate_all()
    self.env['ir.ui.menu'].load_menus(False)

%timeit load_menus()
```
Before previous commit on _visible_menu_ids:
59.7 ms ± 472 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Just before this commit:
46.1 ms ± 140 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
After (also compute more values for the web_load_menu()):
35.3 ms ± 363 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)


## Performance of session_info after this PR

```python
from odoo.http import _request_stack
from odoo.tools import DotDict
from odoo.tools import mute_logger

fake_req = DotDict(dict(
    db=self.env.cr.dbname,
    registry=env.registry,
    uid=env.uid,
    lang=env['res.lang']._get_data(code='en_US'),
    session=DotDict(
        {
            'context': {'lang': 'en_US'},
            'db': self.env.cr.dbname,
            'debug': '',
            'uid': env.uid,
        }
    ),
    cookies={
        "cids": '1'
    }
))
_request_stack.push(fake_req)

def session_info():
    with mute_logger('odoo.registry'):
        self.env.registry.clear_all_caches()
    self.env.invalidate_all()
    self.env['ir.http'].session_info()

%timeit session_info()
```
Before:
70.7 ms ± 603 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
After:
45.5 ms ± 688 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

See https://github.com/odoo/enterprise/pull/74394

task-4360403